+++ app/src/main/java/com/abloom/mery/presentation/ui/signup/MarryDateFragment.kt
@@ -2,12 +2,9 @@ package com.abloom.mery.presentation.ui.signup
 
 import android.os.Bundle
 import android.view.View
-import androidx.navigation.fragment.findNavController
 import com.abloom.mery.R
 import com.abloom.mery.databinding.FragmentMarryDateBinding
 import com.abloom.mery.presentation.common.base.BaseFragment
-import com.abloom.mery.presentation.common.view.setOnActionClick
-import com.abloom.mery.presentation.common.view.setOnNavigationClick
 import dagger.hilt.android.AndroidEntryPoint
 
 
@@ -23,13 +20,6 @@ class MarryDateFragment : BaseFragment<FragmentMarryDateBinding>(R.layout.fragme
 
     private fun initListener() {
 
-        binding.appbarMarryDate.setOnNavigationClick {
-            findNavController().popBackStack()
-        }
-
-        binding.appbarMarryDate.setOnActionClick {
-            //  TODO("날짜 업데이트 후 Signup/Step03(이름정보입력)로 이동 하는 로직 구현")
-        }
 
     }
