+++ app/src/main/res/navigation/signup_nav_graph.xml
@@ -1,0 +1,27 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/signup_nav_graph"
+    app:startDestination="@id/brideGroomSelectionFragment">
+
+    <fragment
+        android:id="@+id/brideGroomSelectionFragment"
+        android:name="com.abloom.mery.presentation.ui.signup.BrideGroomSelectionFragment"
+        android:label="fragment_bride_groom_selection"
+        tools:layout="@layout/fragment_bride_groom_selection">
+        <action
+            android:id="@+id/action_brideGroomSelectionFragment_to_marryDateFragment"
+            app:destination="@id/marryDateFragment" />
+    </fragment>
+
+    <fragment
+        android:id="@+id/marryDateFragment"
+        android:name="com.abloom.mery.presentation.ui.signup.MarryDateFragment"
+        android:label="fragment_marry_date"
+        tools:layout="@layout/fragment_marry_date">
+
+
+    </fragment>
+
+</navigation>
