+++ app/src/main/res/values/strings.xml
@@ -46,4 +46,5 @@
 
     <string name="marry_date_guide_text">결혼예정일을 알려주세요.</string>
     <string name="marry_date_guide_detail_text">결혼까지 남은 기간에 따라 문답을 추천해드릴게요.\n정확한 예정일이 없다면 대략적으로 선택해주세요.</string>
+
 </resources>
